+++ functions/Convert-EventLogRecord.ps1
@@ -1,6 +1,6 @@
 ﻿Function Convert-EventLogRecord {
     [cmdletbinding()]
-    [alias("clr")]
+    [alias('clr')]
 
     Param(
         [Parameter(
@@ -19,7 +19,7 @@
     Process {
         foreach ($record in $LogRecord) {
             Write-Verbose "[PROCESS] Processing event id $($record.ID) from $($record.LogName) log on $(($record.MachineName).ToUpper())"
-            Write-Verbose "[PROCESS] Creating XML data"
+            Write-Verbose '[PROCESS] Creating XML data'
             [xml]$r = $record.ToXml()
 
             #2 June 2024 It is possible the event data has the same name, like ID
@@ -28,40 +28,40 @@
                 LogName     = $record.LogName
                 RecordType  = $record.LevelDisplayName
                 TimeCreated = $record.TimeCreated
-                RecordID          = $record.Id
+                ID          = $record.Id
             }
 
             if ($r.Event.EventData.Data.Count -gt 0) {
-                Write-Verbose "[PROCESS] Parsing event data"
+                Write-Verbose '[PROCESS] Parsing event data'
                 if ($r.Event.EventData.Data -is [array]) {
-                <#
+                    <#
                     I only want to enumerate with the For loop if the data is an array of objects
                     If the data is just a single string like Foo, then when using the For loop,
                     the data value will be the F and not the complete string, Foo.
                 #>
-                for ($i = 0; $i -lt $r.Event.EventData.Data.count; $i++) {
-                    $data = $r.Event.EventData.data[$i]
-                    #test if there is structured data or just text
-                    if ($data.name) {
-                        $Name = $data.name
-                        $Value = $data.'#text'
-                    }
-                    else {
-                        Write-Verbose "[PROCESS] No data property name detected"
-                        $Name = "RawProperties"
-                        #data will likely be an array of strings
-                        [string[]]$Value = $data
-                    }
+                    for ($i = 0; $i -lt $r.Event.EventData.Data.count; $i++) {
+                        $data = $r.Event.EventData.data[$i]
+                        #test if there is structured data or just text
+                        if ($data.name) {
+                            $Name = $data.name
+                            $Value = $data.'#text'
+                        }
+                        else {
+                            Write-Verbose '[PROCESS] No data property name detected'
+                            $Name = 'RawProperties'
+                            #data will likely be an array of strings
+                            [string[]]$Value = $data
+                        }
 
-                    if ($h.Contains("RawProperties")) {
-                        Write-Verbose "[PROCESS] Appending to RawProperties"
-                        $h.RawProperties += $value
-                    }
-                    else {
-                        Write-Verbose "[PROCESS] Adding $name"
-                        $h.add($name, $Value)
-                    }
-                } #for data
+                        if ($h.Contains('RawProperties')) {
+                            Write-Verbose '[PROCESS] Appending to RawProperties'
+                            $h.RawProperties += $value
+                        }
+                        else {
+                            Write-Verbose "[PROCESS] Adding $name"
+                            $h.add($name, $Value)
+                        }
+                    } #for data
                 } #data is an array
                 else {
                     $data = $r.Event.EventData.data
@@ -70,14 +70,14 @@
                         $Value = $data.'#text'
                     }
                     else {
-                        Write-Verbose "[PROCESS] No data property name detected"
-                        $Name = "RawProperties"
+                        Write-Verbose '[PROCESS] No data property name detected'
+                        $Name = 'RawProperties'
                         #data will likely be an array of strings
                         [string[]]$Value = $data
                     }
 
-                    if ($h.Contains("RawProperties")) {
-                        Write-Verbose "[PROCESS] Appending to RawProperties"
+                    if ($h.Contains('RawProperties')) {
+                        Write-Verbose '[PROCESS] Appending to RawProperties'
                         $h.RawProperties += $value
                     }
                     else {
@@ -87,15 +87,15 @@
                 }
             } #if data
             else {
-                Write-Verbose "[PROCESS] No event data to process"
+                Write-Verbose '[PROCESS] No event data to process'
             }
 
-            $h.Add("Message", $record.Message)
-            $h.Add("Keywords", $record.KeywordsDisplayNames)
-            $h.Add("Source", $record.ProviderName)
-            $h.Add("Computername", $record.MachineName)
+            $h.Add('Message', $record.Message)
+            $h.Add('Keywords', $record.KeywordsDisplayNames)
+            $h.Add('Source', $record.ProviderName)
+            $h.Add('Computername', $record.MachineName)
 
-            Write-Verbose "[PROCESS] Creating custom object"
+            Write-Verbose '[PROCESS] Creating custom object'
             New-Object -TypeName PSObject -Property $h
         } #foreach record
     } #process