+++ functions/Get-TypeMember.ps1
@@ -137,3 +137,53 @@ Function Get-TypeMember {
     } #end
 } #close function
 
+Function Get-TypeConstructor {
+    [cmdletbinding(DefaultParameterSetName = 'member')]
+    [OutputType('psTypeMemberConstructor')]
+    [alias("ctor")]
+    Param (
+        [Parameter(
+            Position = 0,
+            Mandatory,
+            HelpMessage = 'Specify a .NET type name like DateTime'
+        )]
+        [ValidateNotNullOrEmpty()]
+        [type]$TypeName
+    )
+
+    Begin {
+        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Starting $($MyInvocation.MyCommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running under PowerShell version $($PSVersionTable.PSVersion)"
+    } #begin
+
+    Process {
+        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Processing $($typename.name)"
+        $Constructors = $typename.GetConstructors()
+        if ($Constructors) {
+            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Found $($Constructors.count) constructors"
+            Foreach ($c in $Constructors) {
+                $cParams = $c.GetParameters()
+                if ($cParams) {
+                    $newParams = $cParams | Select-Object ParameterType,
+                    @{Name="ParameterName";Expression = { $_.Name}}
+                }
+                else {
+                    $newParams = @()
+                }
+
+                [PSCustomObject]@{
+                    PSTypeName = 'psTypeMemberConstructor'
+                    Type       = $typename.FullName
+                    Parameters = $newParams
+                }
+            } #foreach c
+        } #if Constructors found
+        else {
+            Write-Warning "No constructors found for $($typename.name)"
+        }
+    } #process
+
+    End {
+        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Ending $($MyInvocation.MyCommand)"
+    } #end
+} #close function