+++ samples/DiskData.ps1
@@ -19,7 +19,7 @@ Function Get-DiskData {
             $data = Get-CimInstance -Class Win32_logicaldisk -Filter "DriveType=3" -ComputerName $Computername -ErrorAction Stop
             $data | ForEach-Object {
                 Write-Detail "Calculating PctFree for $($_.DeviceID)" -Prefix PROCESS -Time | Write-Verbose
-                $_ | Add-Member -MemberType ScriptProperty -Name PctFree -Value { Format-Percent -Value $this.freespace -Total $this.size -Decimal 2 } -Force
+                $_ | Add-Member -MemberType ScriptProperty -Name PctFree -Value { Format-Percent -Value $this.FreeSpace -Total $this.Size -Decimal 2 } -Force
             }
             $data
         }